+++ src/component/viewport/constants.ts
@@ -1,4 +1,4 @@
-import { createCheckerboardPattern } from "./utils/createCheckerboardPattern";
+import { createCheckerboardPattern } from "./utils/create-checkerboard-pattern";
 
 export const GRID_SIZE = 40;
 export const CLEAR_ON_DESIGN_PATTERN = createCheckerboardPattern(
