+++ src/component/viewport/handlers/mouse-handlers.ts
@@ -1,14 +1,14 @@
 import React, { useRef } from 'react';
 import Konva from "konva";
 import { GRID_SIZE } from "../constants";
-import UndoManager from '../../viewport/utils/undo-manager';
+import UndoManager from '../utils/undo-manager';
 
 Konva.dragButtons = [0, 2];  // Enable dragging with left (0) and right (2) mouse buttons
 
 // Create an instance of UndoManager with a specified limit for undo history
 const undoManager = new UndoManager(100); // Adjust the limit value as needed
 
-export const mouseHandlers = (
+export const useMouseHandlers = (
   onPixelPaint?: (x: number, y: number) => void,
   isEditing?: boolean,
   stageRef?: React.RefObject<Konva.Stage>,
