+++ src/component/viewport/utils/create-checkerboard-pattern.ts
