+++ src/component/viewport/utils/index.ts
@@ -1,3 +1,4 @@
-export * from './createCheckerboardPattern';
-export * from './drawGrid';
-export * from './renderPixelColor';
+export * from './create-checkerboard-pattern';
+export * from './draw-grid';
+export * from './render-pixel-color';
+export * from './undo-manager'
