+++ src/component/viewport/utils/render-pixel-color.ts
